feat(*): add custom entrypoint for wagi executor
resolves #69
