+++ crates/loader/src/local/mod.rs
@@ -7,6 +7,9 @@ mod assets;
 /// Configuration representation for a Spin apoplication as a local spin.toml file.
 mod config;
 
+#[cfg(test)]
+mod tests;
+
 use anyhow::{anyhow, Context, Result};
 use config::{RawAppInformation, RawAppManifest, RawComponentManifest};
 use futures::future;
@@ -119,117 +122,5 @@ fn info(raw: RawAppInformation, src: impl AsRef<Path>) -> ApplicationInformation
         trigger: raw.trigger,
         namespace: raw.namespace,
         origin: ApplicationOrigin::File(src.as_ref().to_path_buf()),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::local::config::RawModuleSource;
-
-    use super::*;
-    use anyhow::Result;
-    use spin_config::{ApplicationTrigger, HttpExecutor, TriggerConfig};
-    use std::path::PathBuf;
-
-    const MANIFEST: &str = "tests/valid-with-files/spin.toml";
-    #[tokio::test]
-    async fn test_from_local_source() -> Result<()> {
-        let dir: Option<PathBuf> = None;
-        let cfg = from_file(MANIFEST, dir).await?;
-
-        assert_eq!(cfg.info.name, "spin-local-source-test");
-        assert_eq!(cfg.info.version, "1.0.0");
-        assert_eq!(cfg.info.api_version, "0.1.0");
-        assert_eq!(
-            cfg.info.authors[0],
-            "Fermyon Engineering <engineering@fermyon.com>"
-        );
-
-        let ApplicationTrigger::Http(http) = cfg.info.trigger;
-        assert_eq!(http.base, "/".to_string());
-
-        let TriggerConfig::Http(http) = cfg.components[0].trigger.clone();
-        assert_eq!(http.executor.unwrap(), HttpExecutor::Spin);
-        assert_eq!(http.route, "/...".to_string());
-
-        assert_eq!(cfg.components[0].wasm.mounts.len(), 1);
-
-        assert_eq!(
-            cfg.info.origin,
-            ApplicationOrigin::File("tests/valid-with-files/spin.toml".into())
-        );
-
-        Ok(())
-    }
-
-    const CFG_TEST: &str = r#"
-    apiVersion  = "0.1.0"
-    name        = "chain-of-command"
-    version     = "6.11.2"
-    description = "A simple application that returns the number of lights"
-    authors     = [ "Gul Madred", "Edward Jellico", "JL" ]
-    trigger     = { type = "http", base   = "/" }
-
-    [[component]]
-        source = "path/to/wasm/file.wasm"
-        id     = "four-lights"
-        files  = ["file.txt", "subdir/another.txt"]
-    [component.trigger]
-        route          = "/lights"
-        executor       = "spin"
-    [component.environment]
-        env1 = "first"
-        env2 = "second"
-    
-    [[component]]
-        id = "abc"
-    [component.source]
-        reference = "bindle reference"
-        parcel    = "parcel"
-    [component.trigger]
-        route = "/test"
-    "#;
-
-    #[test]
-    fn test_manifest() -> Result<()> {
-        let cfg: RawAppManifest = toml::from_str(CFG_TEST)?;
-
-        assert_eq!(cfg.info.name, "chain-of-command");
-        assert_eq!(cfg.info.version, "6.11.2");
-        assert_eq!(
-            cfg.info.description,
-            Some("A simple application that returns the number of lights".to_string())
-        );
-
-        let ApplicationTrigger::Http(http) = cfg.info.trigger;
-        assert_eq!(http.base, "/".to_string());
-
-        assert_eq!(cfg.info.authors.unwrap().len(), 3);
-        assert_eq!(cfg.components[0].id, "four-lights".to_string());
-
-        let TriggerConfig::Http(http) = cfg.components[0].trigger.clone();
-        assert_eq!(http.executor.unwrap(), HttpExecutor::Spin);
-        assert_eq!(http.route, "/lights".to_string());
-
-        let test_component = &cfg.components[0];
-        let test_env = &test_component.wasm.environment.as_ref().unwrap();
-        assert_eq!(test_env.len(), 2);
-        assert_eq!(test_env.get("env1").unwrap(), "first");
-        assert_eq!(test_env.get("env2").unwrap(), "second");
-
-        let test_files = &test_component.wasm.files.as_ref().unwrap();
-        assert_eq!(test_files.len(), 2);
-        assert_eq!(test_files[0], "file.txt");
-        assert_eq!(test_files[1], "subdir/another.txt");
-
-        let b = match cfg.components[1].source.clone() {
-            RawModuleSource::Bindle(b) => b,
-            RawModuleSource::FileReference(_) => panic!("expected bindle source"),
-        };
-
-        assert_eq!(b.reference, "bindle reference".to_string());
-        assert_eq!(b.parcel, "parcel".to_string());
-
-        Ok(())
     }
 }
