+++ crates/loader/tests/valid-manifest.toml
@@ -1,0 +1,25 @@
+apiVersion = "0.1.0"
+authors = ["Gul Madred", "Edward Jellico", "JL"]
+description = "A simple application that returns the number of lights"
+name = "chain-of-command"
+trigger = {type = "http", base = "/"}
+version = "6.11.2"
+
+[[component]]
+files = ["file.txt", "subdir/another.txt"]
+id = "four-lights"
+source = "path/to/wasm/file.wasm"
+[component.trigger]
+executor = {type = "spin"}
+route = "/lights"
+[component.environment]
+env1 = "first"
+env2 = "second"
+
+[[component]]
+id = "abc"
+[component.source]
+parcel = "parcel"
+reference = "bindle reference"
+[component.trigger]
+route = "/test"
