+++ crates/loader/tests/wagi-custom-entrypoint.toml
@@ -1,0 +1,14 @@
+name = "spin-wagi-custom-entrypoint"
+apiVersion = "0.1.0"
+version = "1.0.0"
+authors = [ "Fermyon Engineering <engineering@fermyon.com>" ]
+trigger = { type = "http", base = "/" }
+
+[[component]]
+source = "spin-fs.wasm"
+id = "fs"
+files = ["**/*"]
+
+[component.trigger]
+route = "/hello"
+executor =  { type = "wagi", entrypoint = "custom-entrypoint" }
