+++ templates/spin-http/spin.toml
@@ -1,11 +1,12 @@
-name = "spin-hello-world"
-version = "1.0.0"
+apiVersion = "0.1.0"
+authors = ["Radu Matei <radu@fermyon.com>"]
 description = "A simple application that returns hello and goodbye."
-authors = [ "Radu Matei <radu@fermyon.com>" ]
-trigger = { type = "http", base = "/test" }
+name = "spin-hello-world"
+trigger = {type = "http", base = "/test"}
+version = "1.0.0"
 
 [[component]]
-source = "target/wasm32-wasi/release/spinhelloworld.wasm"
 id = "hello"
+source = "target/wasm32-wasi/release/spinhelloworld.wasm"
 [component.trigger]
 route = "/hello"
